+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-papaparse",
-  "version": "4.1.0",
+  "version": "4.2.0",
   "description": "The fastest in-browser CSV (or delimited text) parser for React. It is full of useful features such as CSVReader, CSVDownloader, readString, jsonToCSV, readRemoteFile, ... etc.",
   "author": "Bunlong <bunlong.van@gmail.com>",
   "license": "MIT",
