+++ supports/create-next-app/package.json
@@ -1,20 +1,23 @@
 {
-  "name": "react-papaparse",
+  "name": "create-next-app",
   "version": "0.1.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
-    "next": "11.1.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-papaparse": "file:../.."
-  },
-  "devDependencies": {
-    "@types/react": "17.0.19",
-    "typescript": "4.4.2"
+    "@types/node": "20.4.5",
+    "@types/react": "18.2.17",
+    "@types/react-dom": "18.2.7",
+    "eslint": "8.45.0",
+    "eslint-config-next": "13.4.12",
+    "next": "13.4.12",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-papaparse": "file:../..",
+    "typescript": "5.1.6"
   }
 }
